refactor(server): address C1 review — deep analysis in-domain, explicit evaluatedAt, signer via module
1. Deep analysis moved behind the agent domain: agent/analysis.ts owns
   the GoPlus/Honeypot scanners, token selection, flag aggregation and
   result construction (deepAnalyze); routes/analyze.ts keeps auth,
   latest-resolution and HTTP serialization only. The operation is now
   part of the future runtime interface.

2. evaluatedAt is an explicit input end-to-end: analyzeRisk takes it as
   a parameter (default new Date() — behaviour identical for callers),
   evaluate.ts threads it through both shapes. Determinism tests now use
   fixed timestamps, and a new hour-boundary test proves the timestamp
   is honoured (out-of-hours evaluation scores higher). The D-milestone
   payload carries the timestamp from day one.

3. The three production signing callers (execution, rejection, draft
   finalize) import getSigningAuthority via agent/index — the domain
   module is the single entry point; C2's guard will enforce it.

44/44 tests green.
